tests/tcg/hexagon: fix undefined signed left shift in brev.c
Cast to uint32_t before left-shifting in sext8() to avoid undefined
behavior when shifting a negative value. The subsequent arithmetic
right shift on the int32_t cast correctly sign-extends the result.
Found with UBSan.

+++ tests/tcg/hexagon/brev.c
@@ -106,7 +106,7 @@ uint32_t bitreverse(uint32_t x)
 
 int32_t sext8(int32_t x)
 {
-    return (x << 24) >> 24;
+    return (int32_t)((uint32_t)x << 24) >> 24;
 }
 
 #define TEST_BREV_LOAD(SZ, TYPE, BUF, SHIFT, EXP) \
